+++ apps/api/src/models/valet-assignments/graphql/valet-assignments.resolver.ts
@@ -1,15 +1,24 @@
-import { Args, Mutation, Query, Resolver } from '@nestjs/graphql'
-import { AllowAuthenticated, GetUser } from 'src/common/auth/auth.decorator'
-import { PrismaService } from 'src/common/prisma/prisma.service'
-import { GetUserType } from 'src/common/types'
-import { CreateValetAssignmentInput } from './dtos/create-valet-assignment.input'
+import {
+  ResolveField,
+  Parent,
+  Resolver,
+  Query,
+  Mutation,
+  Args,
+} from '@nestjs/graphql'
+import { ValetAssignmentsService } from './valet-assignments.service'
+import { ValetAssignment } from './entity/valet-assignment.entity'
 import {
   FindManyValetAssignmentArgs,
   FindUniqueValetAssignmentArgs,
 } from './dtos/find.args'
+import { CreateValetAssignmentInput } from './dtos/create-valet-assignment.input'
 import { UpdateValetAssignmentInput } from './dtos/update-valet-assignment.input'
-import { ValetAssignment } from './entity/valet-assignment.entity'
-import { ValetAssignmentsService } from './valet-assignments.service'
+import { checkRowLevelPermission } from 'src/common/auth/util'
+import { GetUserType } from 'src/common/types'
+import { AllowAuthenticated, GetUser } from 'src/common/auth/auth.decorator'
+import { PrismaService } from 'src/common/prisma/prisma.service'
+import { Valet } from 'src/models/valets/graphql/entity/valet.entity'
 
 @Resolver(() => ValetAssignment)
 export class ValetAssignmentsResolver {
@@ -24,8 +33,7 @@ export class ValetAssignmentsResolver {
     @Args('createValetAssignmentInput') args: CreateValetAssignmentInput,
     @GetUser() user: GetUserType,
   ) {
-    console.log('user', user)
-    // checkRowLevelPermission(user, args.uid)
+    checkRowLevelPermission(user, [args.pickupValetId, args.returnValetId])
     return this.valetAssignmentsService.create(args)
   }
 
@@ -45,11 +53,13 @@ export class ValetAssignmentsResolver {
     @Args('updateValetAssignmentInput') args: UpdateValetAssignmentInput,
     @GetUser() user: GetUserType,
   ) {
-    console.log('user', user)
-    // const valetAssignment = await this.prisma.valetAssignment.findUnique({
-    //   where: { id: args.id },
-    // })
-    // checkRowLevelPermission(user, valetAssignment.uid)
+    const valetAssignment = await this.prisma.valetAssignment.findUnique({
+      where: { bookingId: args.bookingId },
+    })
+    checkRowLevelPermission(user, [
+      valetAssignment.pickupValetId,
+      valetAssignment.returnValetId,
+    ])
     return this.valetAssignmentsService.update(args)
   }
 
@@ -59,9 +69,31 @@ export class ValetAssignmentsResolver {
     @Args() args: FindUniqueValetAssignmentArgs,
     @GetUser() user: GetUserType,
   ) {
-    console.log('user', user)
-    // const valetAssignment = await this.prisma.valetAssignment.findUnique(args)
-    // checkRowLevelPermission(user, valetAssignment.uid)
+    const valetAssignment = await this.prisma.valetAssignment.findUnique(args)
+    checkRowLevelPermission(user, [
+      valetAssignment.pickupValetId,
+      valetAssignment.returnValetId,
+    ])
     return this.valetAssignmentsService.remove(args)
+  }
+
+  @ResolveField(() => Valet, { nullable: true })
+  pickupValet(@Parent() parent: ValetAssignment) {
+    if (!parent.pickupValetId) {
+      return null
+    }
+    return this.prisma.valet.findUnique({
+      where: { uid: parent.pickupValetId },
+    })
+  }
+
+  @ResolveField(() => Valet, { nullable: true })
+  returnValet(@Parent() parent: ValetAssignment) {
+    if (!parent.returnValetId) {
+      return null
+    }
+    return this.prisma.valet.findUnique({
+      where: { uid: parent.returnValetId },
+    })
   }
 }
